backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
requests
mistralai
httpx
PyJWT